+++ lib/semantic_navigation/item.rb
@@ -1,12 +1,19 @@
+require 'semantic_navigation/core/render'
+
 module SemanticNavigation
   class Item
-
-    attr :item_id, :name, :url, :sub_menu
+    
+    include ActionView::Helpers::TagHelper
+    include ActionView::Helpers::UrlHelper 
+    include SemanticNavigation::Core::Render::ItemRender
+    
+    attr :item_id, :name, :url, :sub_menu,
+         :item_classes
 
     def initialize(options, sub_menu = nil)
+      @sub_menu = sub_menu
       options.keys.each do |key|
         self.instance_variable_set("@#{key}",options[key])
-        @sub_menu = sub_menu
       end
     end